manataltest/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
bcrypt==4.2.0
click==8.1.7
Django==5.1.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-multipart==0.0.17
rich==13.9.4
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
